added reddit logging + rename functions to helpers
